Fixed:
* overflows

Added:

* full implementation and use of BigInteger
* full support for nearly any bit length for RSA encryption without failure (will only take a very long time as the prime numbers have to be re-calculated again and again..)

+++ src/math/algorithm/advancedEuclideanAlgorithm/TempResult.java
@@ -1,11 +1,13 @@
 package math.algorithm.advancedEuclideanAlgorithm;
 
+import java.math.BigInteger;
+
 public class TempResult
 {
-    public long e, m, emDivided, emRemainer;
-    public Long a, b;
+    public BigInteger e, m, emDivided, emRemainer;
+    public BigInteger a, b;
 
-    public TempResult(long e, long m)
+    public TempResult(BigInteger e, BigInteger m)
     {
         this.e = e;
         this.m = m;
