+++ src/rsa/PublicKey.java
@@ -4,9 +4,9 @@ import java.math.BigInteger;
 
 public class PublicKey
 {
-    public long n, e;
+    public BigInteger n, e;
 
-    public PublicKey(long n, long e)
+    public PublicKey(BigInteger n, BigInteger e)
     {
         this.n = n;
         this.e = e;
