+++ src/rsa/RSAEncryption.java
@@ -6,11 +6,11 @@ public class RSAEncryption
 {
     public BigInteger encrypt(BigInteger symbol, PublicKey key)
     {
-        return symbol.modPow(BigInteger.valueOf(key.e), BigInteger.valueOf(key.n));
+        return symbol.modPow(key.e, key.n);
     }
 
     public BigInteger decrypt(BigInteger symbol, PrivateKey key)
     {
-        return symbol.modPow(BigInteger.valueOf(key.d), BigInteger.valueOf(key.n));
+        return symbol.modPow(key.d, key.n);
     }
 }
